Added authentication to endpoints

+++ src/modules/rooms/RoomApi.ts
@@ -1,6 +1,8 @@
 import { NextFunction, Request, Response, Router } from 'express';
 import { Container } from 'typedi';
 
+import * as Auth from '../../middlewares/auth.middleware';
+
 import RoomService from './RoomService';
 
 import Random from "../../utils/random";
@@ -25,7 +27,7 @@ async function createRoom(req: Request, res: Response, next: NextFunction): Prom
     const service = Container.get<RoomService>(RoomService);
     const { name } = req.body;
     const id = Random.id();
-    const owner = '6jrHnwiHihRNeWqrKirW'  //TODO: replace when we have authentication
+    const owner = req.user.id;
 
     await service.create(id, name, owner);
 
@@ -41,6 +43,11 @@ async function updateRoom(req: Request, res: Response, next: NextFunction): Prom
   try {
     const service = Container.get<RoomService>(RoomService);
     const { id, name } = req.body;
+    const room = await service.get(id);
+
+    if(room.owner !== req.user.id) {
+      return res.status(401).json({message: 'Not enough privileges'})
+    }
 
     const updated = await service.update(id, name);
 
@@ -56,6 +63,11 @@ async function deleteRoom(req: Request, res: Response, next: NextFunction): Prom
   try {
     const service = Container.get<RoomService>(RoomService);
     const { id } = req.body;
+    const room = await service.get(id);
+
+    if(room.owner !== req.user.id){
+      return res.status(401).json({message: 'Not enough privileges'})
+    }
 
     const deleted = await service.delete(id);
 
@@ -67,9 +79,8 @@ async function deleteRoom(req: Request, res: Response, next: NextFunction): Prom
 }
 
 export default (app: Router): void => {
-  app.post('/room/', createRoom);
-  app.get('/room/:id', getRoom);
-  app.put('/room/', updateRoom);
-  app.delete('/room/', deleteRoom);
-
+  app.post('/room/', Auth.authorize([]), createRoom);
+  app.get('/room/:id', Auth.authorize([]), getRoom);
+  app.put('/room/', Auth.authorize([]), updateRoom);
+  app.delete('/room/', Auth.authorize([]), deleteRoom);
 };
